refactor(frontend): extract reporting api module
